+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/request/TLRequestMessagesDeleteChatUser.java
@@ -10,9 +10,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.readTLObject;
-import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeTLObject;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
@@ -25,7 +25,7 @@ public class TLRequestMessagesDeleteChatUser extends TLMethod<TLAbsUpdates> {
 
     public static final int CONSTRUCTOR_ID = 0xe0611f16;
 
-    protected int chatId;
+    protected long chatId;
 
     protected TLAbsInputUser userId;
 
@@ -34,7 +34,7 @@ public class TLRequestMessagesDeleteChatUser extends TLMethod<TLAbsUpdates> {
     public TLRequestMessagesDeleteChatUser() {
     }
 
-    public TLRequestMessagesDeleteChatUser(int chatId, TLAbsInputUser userId) {
+    public TLRequestMessagesDeleteChatUser(long chatId, TLAbsInputUser userId) {
         this.chatId = chatId;
         this.userId = userId;
     }
@@ -56,14 +56,14 @@ public class TLRequestMessagesDeleteChatUser extends TLMethod<TLAbsUpdates> {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(chatId, stream);
+        writeLong(chatId, stream);
         writeTLObject(userId, stream);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        chatId = readInt(stream);
+        chatId = readLong(stream);
         userId = readTLObject(stream, context, TLAbsInputUser.class, -1);
     }
 
@@ -85,11 +85,11 @@ public class TLRequestMessagesDeleteChatUser extends TLMethod<TLAbsUpdates> {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChatId() {
+    public long getChatId() {
         return chatId;
     }
 
-    public void setChatId(int chatId) {
+    public void setChatId(long chatId) {
         this.chatId = chatId;
     }
