+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/request/TLRequestMessagesDiscardEncryption.java
@@ -9,9 +9,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.readTLObject;
-import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -23,14 +23,14 @@ public class TLRequestMessagesDiscardEncryption extends TLMethod<TLBool> {
 
     public static final int CONSTRUCTOR_ID = 0xedd923c5;
 
-    protected int chatId;
+    protected long chatId;
 
     private final String _constructor = "messages.discardEncryption#edd923c5";
 
     public TLRequestMessagesDiscardEncryption() {
     }
 
-    public TLRequestMessagesDiscardEncryption(int chatId) {
+    public TLRequestMessagesDiscardEncryption(long chatId) {
         this.chatId = chatId;
     }
 
@@ -51,13 +51,13 @@ public class TLRequestMessagesDiscardEncryption extends TLMethod<TLBool> {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(chatId, stream);
+        writeLong(chatId, stream);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        chatId = readInt(stream);
+        chatId = readLong(stream);
     }
 
     @Override
@@ -77,11 +77,11 @@ public class TLRequestMessagesDiscardEncryption extends TLMethod<TLBool> {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChatId() {
+    public long getChatId() {
         return chatId;
     }
 
-    public void setChatId(int chatId) {
+    public void setChatId(long chatId) {
         this.chatId = chatId;
     }
 }
