+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/request/TLRequestMessagesEditChatTitle.java
@@ -9,10 +9,10 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.readLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.readTLObject;
 import static com.github.badoualy.telegram.tl.StreamUtils.readTLString;
-import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeString;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
@@ -26,7 +26,7 @@ public class TLRequestMessagesEditChatTitle extends TLMethod<TLAbsUpdates> {
 
     public static final int CONSTRUCTOR_ID = 0xdc452855;
 
-    protected int chatId;
+    protected long chatId;
 
     protected String title;
 
@@ -35,7 +35,7 @@ public class TLRequestMessagesEditChatTitle extends TLMethod<TLAbsUpdates> {
     public TLRequestMessagesEditChatTitle() {
     }
 
-    public TLRequestMessagesEditChatTitle(int chatId, String title) {
+    public TLRequestMessagesEditChatTitle(long chatId, String title) {
         this.chatId = chatId;
         this.title = title;
     }
@@ -57,14 +57,14 @@ public class TLRequestMessagesEditChatTitle extends TLMethod<TLAbsUpdates> {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(chatId, stream);
+        writeLong(chatId, stream);
         writeString(title, stream);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "SimplifiableConditionalExpression"})
     public void deserializeBody(InputStream stream, TLContext context) throws IOException {
-        chatId = readInt(stream);
+        chatId = readLong(stream);
         title = readTLString(stream);
     }
 
@@ -86,11 +86,11 @@ public class TLRequestMessagesEditChatTitle extends TLMethod<TLAbsUpdates> {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChatId() {
+    public long getChatId() {
         return chatId;
     }
 
-    public void setChatId(int chatId) {
+    public void setChatId(long chatId) {
         this.chatId = chatId;
     }
